fix(tooltips): fixed class/job value not accurate
