+++ libs/data/src/lib/model/lazy-data-keys.ts
@@ -166,6 +166,19 @@ export const LazyDataKeys = [
   'ventures',
   'voyageSources',
   'weathers',
+  'achievementsDatabasePages',
+  'actionsDatabasePages',
+  'fatesDatabasePages',
+  'fishingSpotsDatabasePages',
+  'itemsDatabasePages',
+  'levesDatabasePages',
+  'mobsDatabasePages',
+  'nodesDatabasePages',
+  'npcsDatabasePages',
+  'questsDatabasePages',
+  'questsText',
+  'statusesDatabasePages',
+  'traitsDatabasePages',
   'koAchievementDescriptions',
   'koAchievements',
   'koActionCategories',
@@ -274,18 +287,5 @@ export const LazyDataKeys = [
   'zhTripleTriadRuleDescriptions',
   'zhTripleTriadRules',
   'zhVentures',
-  'zhWeathers',
-  'achievementsDatabasePages',
-  'actionsDatabasePages',
-  'fatesDatabasePages',
-  'fishingSpotsDatabasePages',
-  'itemsDatabasePages',
-  'levesDatabasePages',
-  'mobsDatabasePages',
-  'nodesDatabasePages',
-  'npcsDatabasePages',
-  'questsDatabasePages',
-  'questsText',
-  'statusesDatabasePages',
-  'traitsDatabasePages'
+  'zhWeathers'
 ]
