+++ libs/data/src/lib/model/lazy-data.ts
@@ -125,6 +125,19 @@ import {LazyTripleTriadRule} from './lazy-triple-triad-rule';
 import {LazyVenture} from './lazy-venture';
 import {LazyVoyageSource} from './lazy-voyage-source';
 import {LazyWeather} from './lazy-weather';
+import {LazyAchievementsDatabasePage} from './lazy-achievements-database-page';
+import {LazyActionsDatabasePage} from './lazy-actions-database-page';
+import {LazyFatesDatabasePage} from './lazy-fates-database-page';
+import {LazyFishingSpotsDatabasePage} from './lazy-fishing-spots-database-page';
+import {LazyItemsDatabasePage} from './lazy-items-database-page';
+import {LazyLevesDatabasePage} from './lazy-leves-database-page';
+import {LazyMobsDatabasePage} from './lazy-mobs-database-page';
+import {LazyNodesDatabasePage} from './lazy-nodes-database-page';
+import {LazyNpcsDatabasePage} from './lazy-npcs-database-page';
+import {LazyQuestsDatabasePage} from './lazy-quests-database-page';
+import {LazyQuestsText} from './lazy-quests-text';
+import {LazyStatusesDatabasePage} from './lazy-statuses-database-page';
+import {LazyTraitsDatabasePage} from './lazy-traits-database-page';
 import {LazyKoAirshipVoyage} from './lazy-ko-airship-voyage';
 import {LazyKoFate} from './lazy-ko-fate';
 import {LazyKoGatheringBonus} from './lazy-ko-gathering-bonus';
@@ -137,18 +150,6 @@ import {LazyZhGatheringBonus} from './lazy-zh-gathering-bonus';
 import {LazyZhRecipe} from './lazy-zh-recipe';
 import {LazyZhStatus} from './lazy-zh-status';
 import {LazyZhSubmarineVoyage} from './lazy-zh-submarine-voyage';
-import {LazyAchievementsDatabasePage} from './lazy-achievements-database-page';
-import {LazyActionsDatabasePage} from './lazy-actions-database-page';
-import {LazyFatesDatabasePage} from './lazy-fates-database-page';
-import {LazyFishingSpotsDatabasePage} from './lazy-fishing-spots-database-page';
-import {LazyItemsDatabasePage} from './lazy-items-database-page';
-import {LazyLevesDatabasePage} from './lazy-leves-database-page';
-import {LazyMobsDatabasePage} from './lazy-mobs-database-page';
-import {LazyNodesDatabasePage} from './lazy-nodes-database-page';
-import {LazyNpcsDatabasePage} from './lazy-npcs-database-page';
-import {LazyQuestsDatabasePage} from './lazy-quests-database-page';
-import {LazyStatusesDatabasePage} from './lazy-statuses-database-page';
-import {LazyTraitsDatabasePage} from './lazy-traits-database-page';
 
 export interface LazyData {
   achievementSearch: Array<LazyAchievementSearch>;
@@ -318,6 +319,19 @@ export interface LazyData {
   ventures: Record<number, LazyVenture>;
   voyageSources: Record<number, LazyVoyageSource[]>;
   weathers: Record<number, LazyWeather>;
+  achievementsDatabasePages: Record<number, LazyAchievementsDatabasePage>;
+  actionsDatabasePages: Record<number, LazyActionsDatabasePage>;
+  fatesDatabasePages: Record<number, LazyFatesDatabasePage>;
+  fishingSpotsDatabasePages: Record<number, LazyFishingSpotsDatabasePage>;
+  itemsDatabasePages: Record<number, LazyItemsDatabasePage>;
+  levesDatabasePages: Record<number, LazyLevesDatabasePage>;
+  mobsDatabasePages: Record<number, LazyMobsDatabasePage>;
+  nodesDatabasePages: Record<number, LazyNodesDatabasePage>;
+  npcsDatabasePages: Record<number, LazyNpcsDatabasePage>;
+  questsDatabasePages: Record<number, LazyQuestsDatabasePage>;
+  questsText: Record<number, LazyQuestsText>;
+  statusesDatabasePages: Record<number, LazyStatusesDatabasePage>;
+  traitsDatabasePages: Record<number, LazyTraitsDatabasePage>;
   koAchievementDescriptions: Record<number, {ko: string}>;
   koAchievements: Record<number, {ko: string}>;
   koActionCategories: Record<number, {ko: string}>;
@@ -427,17 +441,4 @@ export interface LazyData {
   zhTripleTriadRules: Record<number, {zh: string}>;
   zhVentures: Record<number, {zh: string}>;
   zhWeathers: Record<number, {zh: string}>;
-  achievementsDatabasePages: Record<number, LazyAchievementsDatabasePage>;
-  actionsDatabasePages: Record<number, LazyActionsDatabasePage>;
-  fatesDatabasePages: Record<number, LazyFatesDatabasePage>;
-  fishingSpotsDatabasePages: Record<number, LazyFishingSpotsDatabasePage>;
-  itemsDatabasePages: Record<number, LazyItemsDatabasePage>;
-  levesDatabasePages: Record<number, LazyLevesDatabasePage>;
-  mobsDatabasePages: Record<number, LazyMobsDatabasePage>;
-  nodesDatabasePages: Record<number, LazyNodesDatabasePage>;
-  npcsDatabasePages: Record<number, LazyNpcsDatabasePage>;
-  questsDatabasePages: Record<number, LazyQuestsDatabasePage>;
-  questsText: Record<number, object>;
-  statusesDatabasePages: Record<number, LazyStatusesDatabasePage>;
-  traitsDatabasePages: Record<number, LazyTraitsDatabasePage>;
 }
