+++ libs/data/src/lib/model/lazy-quests-database-page.ts
@@ -15,8 +15,7 @@ export interface LazyQuestsDatabasePage {
   jobCategory:   number;
   ko?:           string;
   level:         number;
-  name:          Name;
-  next:          Next[];
+  next:          number[];
   npcs:          number[];
   patch:         number;
   repeatable:    boolean;
@@ -24,7 +23,6 @@ export interface LazyQuestsDatabasePage {
   rewards:       Reward[];
   start:         number;
   startingPoint: StartingPoint | null;
-  text:          Text;
   trades?:       Trade[];
   zh?:           string;
 }
@@ -36,42 +34,6 @@ export interface Description {
   ja?: string;
   ko?: string;
   zh?: string;
-}
-
-export interface Name {
-  de?: string;
-  en?: string;
-  fr?: string;
-  ja?: string;
-}
-
-export interface Next {
-  ActionReward:          number;
-  ActorSpawnSeq:         number[];
-  BeastReputationRank:   number;
-  ClassJobCategory0:     number;
-  ClassJobLevel0:        number;
-  GilReward:             number;
-  GrandCompany:          number;
-  Id:                    string;
-  Id_de:                 string;
-  Id_en:                 string;
-  Id_fr:                 string;
-  Id_ja:                 string;
-  InstanceContentUnlock: number;
-  IsRepeatable:          boolean;
-  IssuerStart:           number;
-  JournalGenre:          number;
-  PreviousQuest:         number[];
-  ReputationReward:      number;
-  TargetEnd:             number;
-  __sheet:               Sheet;
-  index:                 number;
-  subIndex:              number;
-}
-
-export enum Sheet {
-  Quest = "Quest",
 }
 
 export interface Reward {
@@ -94,23 +56,6 @@ export interface StartingPoint {
   y:      number;
   z:      number;
   zoneid: number;
-}
-
-export interface Text {
-  Dialogue?: Dialogue[];
-  Journal?:  Journal[];
-  ToDo?:     Journal[];
-}
-
-export interface Dialogue {
-  npc:   number | null;
-  order: number;
-  text:  Name;
-}
-
-export interface Journal {
-  order: number;
-  text:  Name;
 }
 
 export interface Trade {
